+++ src/app/events/UserUpdate.ts
@@ -14,7 +14,7 @@ export default class implements DiscordEvent {
 			if (newUser.partial) {
 				newUser = await newUser.fetch();
 			}
-			
+		
 			if (oldUser.username !== newUser.username) {
 				const guildMember = await ServiceUtils.getGuildMemberFromUser(newUser as User, process.env.DISCORD_SERVER_ID);